+++ packages/core/src/types/index.ts
@@ -26,6 +26,8 @@ import type {
   ViewportInput,
   RenderingEngineResizeOptions,
 } from './IViewport';
+import type IGenericViewport from './IGenericViewport';
+import type { ViewportContentMode } from './IGenericViewport';
 import type {
   VolumeActor,
   Actor,
@@ -62,6 +64,7 @@ import type {
   Scaling,
   ScalingParameters,
 } from './ScalingParameters';
+import type { ScalingDescriptor } from './ScalingDescriptor';
 import type StackViewportProperties from './StackViewportProperties';
 import type VolumeViewportProperties from './VolumeViewportProperties';
 import type IViewportId from './IViewportId';
@@ -189,6 +192,7 @@ export type {
   IDynamicImageVolume,
   IRenderingEngine,
   ScalingParameters,
+  ScalingDescriptor,
   PTScaling,
   IPointsManager,
   PolyDataPointConfiguration,
@@ -205,6 +209,8 @@ export type {
   IRegisterImageLoader,
   IStreamingVolumeProperties,
   IViewport,
+  IGenericViewport,
+  ViewportContentMode,
   ViewReference,
   DataSetOptions as ImageSetOptions,
   ViewPresentation,